chore: add close modal into its own thing
